+++ src/main/java/insane96mcp/enhancedai/modules/witch/WitchModule.java
@@ -1,5 +1,6 @@
 package insane96mcp.enhancedai.modules.witch;
 
+import insane96mcp.enhancedai.modules.witch.feature.ThirstyWitches;
 import insane96mcp.enhancedai.modules.witch.feature.WitchFleeTarget;
 import insane96mcp.enhancedai.modules.witch.feature.WitchPotionThrowing;
 import insane96mcp.enhancedai.setup.Config;
@@ -11,12 +12,14 @@ public class WitchModule extends Module {
 
 	public WitchPotionThrowing witchPotionThrowing;
 	public WitchFleeTarget witchFleeTarget;
+	public ThirstyWitches thirstyWitches;
 
 	public WitchModule() {
 		super(Config.builder);
 		this.pushConfig(Config.builder);
 		witchPotionThrowing = new WitchPotionThrowing(this);
 		witchFleeTarget = new WitchFleeTarget(this);
+		thirstyWitches = new ThirstyWitches(this);
 		Config.builder.pop();
 	}
 
@@ -25,5 +28,6 @@ public class WitchModule extends Module {
 		super.loadConfig();
 		witchPotionThrowing.loadConfig();
 		witchFleeTarget.loadConfig();
+		thirstyWitches.loadConfig();
 	}
 }